Project_Base/requirements.txt
fastapi
uvicorn
sqlalchemy
paho-mqtt
pydantic